+++ CHANGELOG.md
@@ -2,6 +2,65 @@
 
 <!--next-version-placeholder-->
 
+## v0.3.5 (2023-02-14)
+### Fix
+* Pypi not uploading ([`543d0a0`](https://github.com/Lash-L/southern_company_api/commit/543d0a057f11e7f419c2e43293d552afb25321e6))
+
+## v0.3.4 (2023-02-14)
+### Fix
+* Get accounts test ([`fd8e78e`](https://github.com/Lash-L/southern_company_api/commit/fd8e78e8456b3c2046cc6e9282bcfc75dfd036a7))
+* Forgot await ([`77e0582`](https://github.com/Lash-L/southern_company_api/commit/77e0582309a9ed4547a49bd95622977c74720205))
+
+## v0.3.3 (2023-02-13)
+### Fix
+* Get_accounts to use property ([`0985457`](https://github.com/Lash-L/southern_company_api/commit/09854576d071ec4af761321d9b5c2858a86efe4c))
+
+## v0.3.2 (2023-02-13)
+### Fix
+* Jwt will refresh after 3 hours ([`a16f442`](https://github.com/Lash-L/southern_company_api/commit/a16f442681748cd4df57b7391d1d0f513527d54a))
+
+## v0.3.1 (2023-02-11)
+### Fix
+* Remove optional from properties ([`b6cc793`](https://github.com/Lash-L/southern_company_api/commit/b6cc7935547202801d92c8f3b68da640df65857a))
+
+## v0.3.0 (2023-02-10)
+### Feature
+* Made tokens async props ([`a56487b`](https://github.com/Lash-L/southern_company_api/commit/a56487b4abd7a0668a3a735f125711e966cdfdf6))
+
+### Fix
+* Removed delayed data ([`a0f1f36`](https://github.com/Lash-L/southern_company_api/commit/a0f1f363816eec23f07c00e90592d0ddab946a90))
+
+## v0.2.4 (2023-02-08)
+### Fix
+* Add timezone to datetime ([`2578d5f`](https://github.com/Lash-L/southern_company_api/commit/2578d5f80562f46bad40aa098aadabed88646df9))
+
+## v0.2.3 (2023-02-08)
+### Fix
+* Made account not optional ([`22e6c8b`](https://github.com/Lash-L/southern_company_api/commit/22e6c8b0d58de96e00eb799a68ce34d875499a77))
+
+## v0.2.2 (2023-02-08)
+### Fix
+* Made accounts private ([`33a812a`](https://github.com/Lash-L/southern_company_api/commit/33a812ae767f2f654a999652fcfd6d1538804477))
+
+## v0.2.1 (2023-02-08)
+### Fix
+* Set accounts to member variable ([`9cb81f8`](https://github.com/Lash-L/southern_company_api/commit/9cb81f86e034cdd718fc8cd5feacecc642bb3795))
+
+## v0.2.0 (2023-02-08)
+### Feature
+* Add monthly usage ([`7ae4a6c`](https://github.com/Lash-L/southern_company_api/commit/7ae4a6c7a00190ff91378723170a517439a69212))
+
+## v0.1.1 (2023-02-08)
+### Fix
+* Account hourly time ([`3cad045`](https://github.com/Lash-L/southern_company_api/commit/3cad0451a3cc23e5e3e5ddc8bd1187aeefd5d7a7))
+
+## v0.1.0 (2023-02-08)
+### Feature
+* Added new tests ([`501ba1a`](https://github.com/Lash-L/southern_company_api/commit/501ba1a67bd5cd3459bafcc9f65beac547f7fbd5))
+
+### Fix
+* Resolved issue with sc_token ([`9138c10`](https://github.com/Lash-L/southern_company_api/commit/9138c10610f359fdaa9fe712fdd870ee7d8cf6e9))
+
 ## v0.0.4 (2023-02-07)
 ### Fix
 * Lint for ci ([`286bc2d`](https://github.com/Lash-L/southern_company_api/commit/286bc2d7474da1d9478e63828b137d439e0dccc1))